#81 reorganize and utilize old existing code for checkmate logic handling
